+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 2.1.0
+- macOS: Updated to FreSwift 4.3.0
+- WIN: Updated to FreSharp 2.4.0
+
 ### 2.0.0
 - WIN Convert to C#
 - WIN add FreSharp as a dependency
@@ -10,7 +14,7 @@
 
 ### 1.7.0
 - Upgraded to AIR 32.0.0.116
-- iOS: Updated to FreSwift 3.1.0
+- macOS: Updated to FreSwift 3.1.0
 
 ### 1.6.0
 - Updated to FreSwift 3.0.0